"chore: renaming and providing obvious name for tmp dir"
