+++ __tests__/sandbox.ava.ts
@@ -2,8 +2,8 @@ import test from 'ava';
 import * as net from 'net';
 import { existsSync } from "fs";
 import { writeFile } from 'fs/promises';
-import { Sandbox } from '../src/server/Sandbox';
-import { GenesisAccount, SandboxConfig } from '../src/server/config';
+import { Sandbox } from '../src/sandbox/Sandbox';
+import { GenesisAccount, SandboxConfig } from '../src/sandbox/config';
 import { tmpdir } from 'os';
 import { join } from 'path';
 import { lock } from 'proper-lockfile';
